+++ app/src/main/java/org/simple/clinic/registration/facility/RegistrationFacilitySelectionScreenKey.kt
@@ -4,9 +4,12 @@ import kotlinx.android.parcel.IgnoredOnParcel
 import kotlinx.android.parcel.Parcelize
 import org.simple.clinic.R
 import org.simple.clinic.router.screen.FullScreenKey
+import org.simple.clinic.user.OngoingRegistrationEntry
 
 @Parcelize
-class RegistrationFacilitySelectionScreenKey : FullScreenKey {
+data class RegistrationFacilitySelectionScreenKey(
+    val ongoingRegistrationEntry: OngoingRegistrationEntry
+) : FullScreenKey {
 
   @IgnoredOnParcel
   override val analyticsName = "Registration Facility Selection"