+++ app/src/main/java/org/simple/clinic/registration/facility/RegistrationFacilitySelectionUi.kt
@@ -2,7 +2,6 @@ package org.simple.clinic.registration.facility
 
 import org.simple.clinic.facility.change.FacilitiesUpdateType
 import org.simple.clinic.facility.change.FacilityListItem
-import java.util.UUID
 
 interface RegistrationFacilitySelectionUi: RegistrationFacilitySelectionUiActions {
   fun showProgressIndicator()
@@ -10,6 +9,4 @@ interface RegistrationFacilitySelectionUi: RegistrationFacilitySelectionUiAction
   fun showToolbarWithSearchField()
   fun showToolbarWithoutSearchField()
   fun updateFacilities(facilityItems: List<FacilityListItem>, updateType: FacilitiesUpdateType)
-  fun openIntroVideoScreen()
-  fun showConfirmFacilitySheet(facilityUuid: UUID, facilityName: String)
 }