+++ app/src/main/java/org/simple/clinic/registration/location/RegistrationLocationPermissionEffect.kt
@@ -1,5 +1,7 @@
 package org.simple.clinic.registration.location
 
+import org.simple.clinic.user.OngoingRegistrationEntry
+
 sealed class RegistrationLocationPermissionEffect
 
-object OpenFacilitySelectionScreen: RegistrationLocationPermissionEffect()
+data class OpenFacilitySelectionScreen(val entry: OngoingRegistrationEntry): RegistrationLocationPermissionEffect()