+++ app/src/main/java/org/simple/clinic/registration/location/RegistrationLocationPermissionEffectHandler.kt
@@ -19,7 +19,7 @@ class RegistrationLocationPermissionEffectHandler @AssistedInject constructor(
   fun build(): ObservableTransformer<RegistrationLocationPermissionEffect, RegistrationLocationPermissionEvent> {
     return RxMobius
         .subtypeEffectHandler<RegistrationLocationPermissionEffect, RegistrationLocationPermissionEvent>()
-        .addAction(OpenFacilitySelectionScreen::class.java, uiActions::openFacilitySelectionScreen, schedulers.ui())
+        .addConsumer(OpenFacilitySelectionScreen::class.java, { uiActions.openFacilitySelectionScreen(it.entry) }, schedulers.ui())
         .build()
   }
 }