+++ app/src/main/java/org/simple/clinic/registration/location/RegistrationLocationPermissionModel.kt
@@ -2,11 +2,16 @@ package org.simple.clinic.registration.location
 
 import android.os.Parcelable
 import kotlinx.android.parcel.Parcelize
+import org.simple.clinic.user.OngoingRegistrationEntry
 
 @Parcelize
-class RegistrationLocationPermissionModel: Parcelable {
+data class RegistrationLocationPermissionModel(
+    val ongoingRegistrationEntry: OngoingRegistrationEntry
+): Parcelable {
 
   companion object {
-    fun create(): RegistrationLocationPermissionModel = RegistrationLocationPermissionModel()
+    fun create(ongoingRegistrationEntry: OngoingRegistrationEntry): RegistrationLocationPermissionModel {
+      return RegistrationLocationPermissionModel(ongoingRegistrationEntry)
+    }
   }
 }